improve target selection after committing editor
app.select_target doesn't do the right thing when an image appears multiple
times in the tree due to grouping by a set field. It always selects the first
instance of the image which may not be the one we want.

This patch updates the editor to construct the desired target path to use
once the tree has been rebuilt post-edit